move scroll wrapper styled component to library
