+++ src/renderer/screens/Import/Ledger/Manage.tsx
@@ -16,10 +16,10 @@ import { getSortedLocalLedgerAddresses } from '@/renderer/utils/ImportUtils';
 import { useAddresses } from '@/renderer/contexts/import/Addresses';
 import { useState } from 'react';
 import { ButtonPrimaryInvert } from '@/renderer/kits/Buttons/ButtonPrimaryInvert';
+import { Scrollable } from '@/renderer/library/utils';
 import {
   ControlsWrapper,
   StatsFooter,
-  Scrollable,
   SortControlLabel,
 } from '@/renderer/utils/common';
 import type { ImportLedgerManageProps } from '../types';
